backport refactoring from proposed rustgen generator: improvements to the TemplateModel, make imports handling generic. from @sneakers-the-rat
