+++ linkml/generators/common/build.py
@@ -4,39 +4,24 @@ Models for intermediate build results
 (see PydanticGenerator for example implementation and use)
 """
 
-import dataclasses
-from typing import Annotated, Any, TypeVar
+from typing import Annotated, TypeVar
 
 from linkml_runtime.linkml_model import (
     ClassDefinition,
+    Definition,
     EnumDefinition,
     SchemaDefinition,
     SlotDefinition,
     TypeDefinition,
 )
-from pydantic import BaseModel, ConfigDict, GetCoreSchemaHandler
+from pydantic import BaseModel, ConfigDict, GetCoreSchemaHandler, GetPydanticSchema
 from pydantic_core import core_schema
 
 T = TypeVar("T", bound="BuildResult", covariant=True)
+Tsv = TypeVar("Tsv", bound=Definition, covariant=True)
 
 
-@dataclasses.dataclass()
-class SkipValidation:
-    """
-    A version of :class:`pydantic.SkipValidation` that actually skips generating the
-    schema for the field entirely - useful for including types that don't need to be validated
-    like the metamodel dataclasses
-    """
-
-    def __class_getitem__(cls, item: Any) -> Any:
-        return Annotated[item, SkipValidation()]
-
-    @classmethod
-    def __get_pydantic_core_schema__(cls, source: Any, handler: GetCoreSchemaHandler) -> core_schema.CoreSchema:
-        return core_schema.any_schema()
-
-    __hash__ = object.__hash__
-
+SkipValidation = Annotated[Tsv, GetPydanticSchema(lambda tp, handler: core_schema.any_schema())]
 
 class BuildResult(BaseModel):
     """
